+++ tests/mocks.js
@@ -16,3 +16,12 @@ export const mockLogger = () => {
     },
   }));
 };
+
+/**
+ * Mocks the config module and all of the commonly used functions.
+ */
+export const mockConfig = () => {
+  jest.unstable_mockModule('../../src/config.js', () => ({
+    getConfigValue: jest.fn(),
+  }));
+};
